Revert "Merge pull request #2336 from berryzplus/feature/clang-cl"
This reverts commit 7e5552ed10a4946c8b320a8ae6f2b4b64b3c0ffd, reversing
changes made to 6aff8f67c73cb9c9e81bde9adcf058bcfdaa6201.

+++ sakura_core/util/window.cpp
@@ -338,7 +338,7 @@ HFONT GetSystemBasedFont( LONG nLogicalHeight )
 	// キー:文字列化したLOGFONT
 	static std::map<std::wstring, HFONT> fontStock;
 
-	NONCLIENTMETRICS metrics = { UINT(CCSIZEOF_STRUCT(NONCLIENTMETRICS, lfMessageFont)) };
+	NONCLIENTMETRICS metrics = { CCSIZEOF_STRUCT( NONCLIENTMETRICS, lfMessageFont ) };
 	if( !SystemParametersInfo( SPI_GETNONCLIENTMETRICS, 0, &metrics, 0 ) ) {
 		return nullptr;
 	}
